requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nibabel==5.3.3
numpy==2.0.2
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
streamlit==1.49.1
typing_extensions==4.15.0
zipp==3.23.0
